Move notification related functions from db_block.cpp to db_notify.cpp
