+++ libraries/chain/database.cpp
@@ -32,3 +32,4 @@
 #include "db_market.cpp"
 #include "db_update.cpp"
 #include "db_witness_schedule.cpp"
+#include "db_notify.cpp"
